pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Агент для автоматизации разработки на GitHub"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "pygithub>=2.1.0",
    "gitpython>=3.1.40",
    "google-genai>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
code-agent = "coding_agent.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
